% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspect_importance.R
\name{get_sample}
\alias{get_sample}
\title{Function for getting binary matrix}
\usage{
get_sample(n, p, sample_method = c("default", "binom"), f = 2)
}
\arguments{
\item{n}{number of rows}

\item{p}{number of columns}

\item{sample_method}{sampling method}

\item{f}{frequency for binomial sampling}
}
\value{
a binary matrix
}
\description{
Function creates binary matrix, to be used in aspect_importance method. It
starts with a zero matrix. Then it replaces some zeros with ones. If
\code{sample_method = "default"} it randomly replaces one or two zeros per
row. If \code{sample_method = "binom"} it replaces random number of zeros
per row - average number of replaced zeros can be controlled by parameter
\code{sample_method = "f"}. Function doesn't allow the returned matrix to
have rows with only zeros.
}
\examples{
 get_sample(100,6,"binom",3)

}
