% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_variables.R
\name{list_variables}
\alias{list_variables}
\title{Cuts tree at custom height and returns a list}
\usage{
list_variables(x, h)
}
\arguments{
\item{x}{hclust object}

\item{h}{correlation value for tree cutting}
}
\value{
list of aspects
}
\description{
This function creates aspect list after cutting a tree at a given height.
}
\examples{
library("DALEX")
dragons_data <- dragons[,c(2,3,4,7,8)]
cv <- cluster_variables(dragons_data, clust_method = "complete")
list_variables(cv, h = 0.5)

}
