% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_variables.R
\name{cluster_variables}
\alias{cluster_variables}
\alias{cluster_variables.default}
\title{Groups numeric features into aspects}
\usage{
cluster_variables(x, ...)

\method{cluster_variables}{default}(x, clust_method = "complete", ...)
}
\arguments{
\item{x}{dataframe with only numeric columns}

\item{...}{other parameters}

\item{clust_method}{the agglomeration method to be used
see \code{\link[stats]{hclust}} methods}
}
\value{
list of aspects
}
\description{
Divides correlated features into groups, called aspects. Division is based on
correlation cutoff level.
}
\examples{
library("DALEX")
dragons_data <- dragons[,c(2,3,4,7,8)]
cluster_variables(dragons_data, clust_method = "complete")

}
