% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_importance.R
\name{hierarchical_importance}
\alias{hierarchical_importance}
\alias{plot.hierarchical_importance}
\title{Calculates importance of hierarchically grouped aspects}
\usage{
hierarchical_importance(
  x,
  data,
  y = NULL,
  predict_function = predict,
  type = "predict",
  new_observation = NULL,
  N = 1000,
  loss_function = DALEX::loss_root_mean_square,
  B = 10,
  clust_method = "complete",
  ...
)

\method{plot}{hierarchical_importance}(
  x,
  absolute_value = FALSE,
  show_labels = TRUE,
  add_last_group = TRUE,
  axis_lab_size = 10,
  text_size = 3,
  ...
)
}
\arguments{
\item{x}{a model to be explained}

\item{data}{dataset, should be without target variable}

\item{y}{true labels for \code{data}, will be extracted from \code{x}
if it's an explainer,
  need to be provided if feature importance is to be calculated}

\item{predict_function}{predict function}

\item{type}{if \code{predict} aspect_importance is used, if
\code{model} is used, than feature_importance is calculated}

\item{new_observation}{selected observation with columns that corresponds to
variables used in the model (it should be without target variable),
if NULL feature_importance will be calculated}

\item{N}{number of observations to be sampled (with replacement) from data}

\item{loss_function}{a function that will be used to assess variable importance}

\item{B}{integer, number of permutation rounds to perform on each variable in 
feature importance calculation. By default it's \code{10}.}

\item{clust_method}{the agglomeration method to be used, see
\code{\link[stats]{hclust}} methods}

\item{...}{other parameters}

\item{absolute_value}{if TRUE, aspect importance values will be drawn as
absolute values}

\item{show_labels}{if TRUE, plot will have annotated axis Y}

\item{add_last_group}{if TRUE and \code{type} is \code{predict}, plot will 
draw connecting line between last two groups at the level of 105% of the 
biggest importance value, for \code{model} this line is always drawn}

\item{axis_lab_size}{size of labels on axis Y, if applicable}

\item{text_size}{size of labels annotating values of aspects importance}
}
\value{
ggplot
}
\description{
This function plots tree that shows order of feature grouping and aspect
importance or feature importance values of every newly created aspect.
}
\examples{
library(DALEX)
apartments_num <- apartments[,unlist(lapply(apartments, is.numeric))]
apartments_num_lm_model <- lm(m2.price ~ ., data = apartments_num)
hi <- hierarchical_importance(x = apartments_num_lm_model,
 data = apartments_num[,-1],
 y = apartments_num[,1],
 type = "model")
plot(hi, add_last_group = TRUE, absolute_value = TRUE)

}
