% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triplot.R
\name{plot.triplot}
\alias{plot.triplot}
\title{Plots triplot with correlation values}
\usage{
\method{plot}{triplot}(x, ...)
}
\arguments{
\item{x}{triplot object}

\item{...}{other parameters}
}
\value{
plot
}
\description{
Plots triplot that sum up automatic aspect/feature importance grouping
}
\examples{
library(DALEX)
set.seed(123)
apartments_num <- apartments[,unlist(lapply(apartments, is.numeric))]
apartments_num_lm_model <- lm(m2.price ~ ., data = apartments_num)
apartments_num_new_observation <- apartments_num[30, ]
explainer_apartments <- explain(model = apartments_num_lm_model,
 data = apartments_num[,-1],
 y = apartments_num[, 1])
apartments_tri <- calculate_triplot(x = explainer_apartments,
 new_observation = apartments_num_new_observation[-1])
plot(apartments_tri)

}
