% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triplot.R
\name{plot.triplot}
\alias{plot.triplot}
\title{Plots triplot with correlation values}
\usage{
\method{plot}{triplot}(x, ...)
}
\arguments{
\item{x}{triplot object}

\item{...}{other parameters}
}
\value{
plot
}
\description{
Plots triplot that sum up automatic aspect/feature importance grouping
}
\examples{
library(DALEX)
library(ingredients)
apartments_num <- apartments[,unlist(lapply(apartments, is.numeric))]
apartments_num_lm_model <- lm(m2.price ~ ., data = apartments_num)
apartments_num_new_observation <- apartments_num[30,-1]
apartments_num_mod <- apartments_num[,-1]
tri <- calculate_triplot(x = apartments_num_lm_model,
  data = apartments_num_mod,
  new_observation = apartments_num_new_observation,
  add_importance_labels = FALSE)

 plot(tri)

}
