% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspect_importance.R
\name{print.aspect_importance}
\alias{print.aspect_importance}
\title{Function for printing aspect_importance results}
\usage{
\method{print}{aspect_importance}(x, show_features = FALSE, show_corr = FALSE, ...)
}
\arguments{
\item{x}{object of aspect_importance class}

\item{show_features}{show list of features for every aspect}

\item{show_corr}{show if all features in aspect are pairwise positively
correlated, (works for numeric features)}
}
\description{
This function prints the results of aspect_importance.
}
